+++ src/class_base.h
@@ -1120,7 +1120,7 @@ void CLASS_BASE_DECL_SPEC wait( unsigned long ms );
 
 std::string CLASS_BASE_DECL_SPEC get_uuid( );
 
-size_t CLASS_BASE_DECL_SPEC get_random( );
+uint32_t CLASS_BASE_DECL_SPEC get_random( );
 
 std::string CLASS_BASE_DECL_SPEC get_random_hash( );
